requirements.txt
streamlit
openai
pandas
PyYAML